Migrate tests from webhooks to root package
